+++ TDTChocolate/FoundationAdditions/NSDictionary+TDTAdditions.m
@@ -14,4 +14,17 @@
   return [dict copy];
 }
 
+- (BOOL)tdt_writeInBinaryFormatToFile:(NSString *)path
+                           atomically:(BOOL)useAuxiliaryFile {
+  NSData *data = [NSPropertyListSerialization dataWithPropertyList:self
+                                                            format:NSPropertyListBinaryFormat_v1_0
+                                                           options:0
+                                                             error:NULL];
+  if (!data) {
+    return NO;
+  }
+
+  return [data writeToFile:path atomically:useAuxiliaryFile];
+}
+
 @end